+++ playground/lib/src/presentation/editor/widgets/panels/widgets/tree_list_panel/widgets/drop_between_nodes_area.dart
@@ -12,9 +12,10 @@ class TreeListDropBetweenNodesArea extends StatefulWidget {
   });
 
   final CNode node;
-  final int index;
+  final double index;
   final Function(
-      DragTargetMoveSingleNodeModel event, CNode newParent, int index) onMove;
+          DragTargetMoveSingleNodeModel event, CNode newParent, double index)
+      onMove;
   final Function(CNode newNode, CNode parent) onAdd;
 
   @override
